pyspark-sdk/pyproject.toml
[build-system]
requires = ["setuptools<82", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sagemaker_feature_store_pyspark"
description = "Amazon SageMaker FeatureStore PySpark Bindings"
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.8"
dynamic = ["version"]
keywords = ["ML", "Amazon", "AWS", "AI", "FeatureStore", "SageMaker"]
authors = [
    { name = "Amazon Web Services" },
]
dependencies = [
    "setuptools<82",
]

[project.optional-dependencies]
unit = [
    "numpy",
    "pytest",
    "pytest-cov",
    "pytest-rerunfailures",
    "pytest-xdist",
    "coverage",
]
integ = [
    "boto3",
    "numpy",
    "sagemaker-mlops>=1.8",
    "sagemaker-core>=2.8",
    "pyiceberg[glue,pyarrow,pandas]>=0.9.0",
]
lint = [
    "flake8>=6.0.0",
]
release = [
    "twine>=1.12.0",
]

[tool.pytest.ini_options]
addopts = "-vv -s"
testpaths = ["tests"]

[tool.flake8]
exclude = [".git", "__pycache__", "build", "dist", ".eggs"]
max-complexity = 10
max-line-length = 120
